Fixing database - creating candies route in server
