+++ server/db/models/candy.js
@@ -1,35 +1,35 @@
 const Sequelize = require('sequelize')
-const db = require('./database')
+const db = require('../db')
 
-const Candy = db.define('candys', {
+const Candy = db.define('candy', {
   name: {
     type: Sequelize.STRING,
     allowNull: false,
     validate: {
-      notEmpty: true,
-    },
+      notEmpty: true
+    }
   },
 
   price: {
     type: Sequelize.FLOAT,
     allowNull: false,
     validate: {
-      notEmpty: true,
-    },
+      notEmpty: true
+    }
   },
   imageUrl: {
-    type: Sequelize.STRING,
+    type: Sequelize.STRING
   },
   description: {
-    type: Sequelize.TEXT,
+    type: Sequelize.TEXT
   },
   quantity: {
     type: Sequelize.INTEGER,
     allowNull: false,
     validate: {
-      notEmpty: true,
-    },
-  },
+      notEmpty: true
+    }
+  }
 })
 
 module.exports = Candy
